Update protocol definitions
