+++ Sources/Swiftagram/Models/Specialized/Specialized.swift
@@ -1,5 +1,5 @@
 //
-//  ResponseType.swift
+//  Specialized.swift
 //  Swiftagram
 //
 //  Created by Stefano Bertagno on 26/08/20.
@@ -8,14 +8,14 @@
 import Foundation
 
 /// A `protocol` describing a generic response returning an element of `Response`.
-public protocol ResponseType: Wrapped {
+public protocol Specialized: Wrapped {
     /// An optional `ResponseError` message returned by a response.
     /// Default emplementation returns failing description, if it exists,
     /// otherwise `.unknown` if `status` is not `ok`, and `nil` if it is.
     var error: ResponseError? { get }
 }
 
-public extension ResponseType {
+public extension Specialized {
     /// The response status.
     @available(*, deprecated, message: "check for `error` instead")
     var status: String! { self["status"].string() }
