--- Tests/SwiftagramTests/Shared/ReferenceType.swift
@@ -1,14 +1,0 @@
-//
-//  ReferenceType.swift
-//  SwiftagramTests
-//
-//  Created by Stefano Bertagno on 28/10/20.
-//
-
-import Foundation
-
-/// A custom reference typed wrapper.
-final class ReferenceType<T> {
-    /// The underlying value.
-    var value: T?
-}
