+++ Tests/SwiftagramTests/Shared/Reflection/Sticker.swift
